added proper migration

+++ db/migrate/20180129034322_change_media_user_id_to_profile_id.rb
@@ -1,6 +1,6 @@
 class ChangeMediaUserIdToProfileId < ActiveRecord::Migration[5.1]
   def change
-    add_column :media, :profile_id, :integer
+    add_column :media, :profile_id, :bigint
     media_type_to_role = {
       "photo" => "Photographer",
       "illustration" => "Illustrator",
